requirements.txt
Django==2.2.1
djangorestframework==3.9.4
requests==2.22.0
